File reorganization, initial aggressiveness rework, added officer/worker labels
