+++ modules/action_notifications.py
@@ -1,3 +1,5 @@
+#Contains functionality for multi-step notifications
+
 from .labels import label
 from .images import free_image
 from .notifications import notification
