+++ modules/scaling.py
@@ -1,3 +1,5 @@
+#Contains functions that scale coordinates and lengths/widths to different resolutions
+
 def scale_coordinates(x, y, global_manager):
     '''
     Description:
