+++ modules/vehicles.py
@@ -1,3 +1,5 @@
+#Contains functionality for vehicle units
+
 from . import text_tools
 from . import utility
 from . import actor_utility
